Fixed bug opening the arg file multiple times when switching views.
